Refined manifest, fix background script TypeError & bugs
