--- chrome-extension-overlay/manifest.json
@@ -1,51 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "Real-time Subtitle Generator & Translator",
-  "version": "1.0.0",
-  "description": "Generate real-time subtitles and translate them using Microsoft Azure Speech Services",
-  "permissions": [
-    "storage",
-    "activeTab",
-    "scripting"
-  ],
-  "background": {
-    "service_worker": "background/background.js",
-    "type": "module"
-  },
-  "content_scripts": [
-    {
-      "matches": ["*://*.youtube.com/*"],
-      "js": ["content/youtube.js", "content/overlay.js"],
-      "run_at": "document_end",
-      "all_frames": false
-    },
-    {
-      "matches": ["*://*.teams.microsoft.com/*", "*://*.teams.live.com/*"],
-      "js": ["content/teams.js", "content/overlay.js"],
-      "run_at": "document_end",
-      "all_frames": false
-    },
-    {
-      "matches": ["*://*.zoom.us/*"],
-      "js": ["content/zoom.js", "content/overlay.js"],
-      "run_at": "document_end",
-      "all_frames": false
-    }
-  ],
-  "action": {
-    "default_popup": "popup/popup.html",
-    "default_title": "Subtitle Generator",
-    "default_icon": {
-      "16": "assets/icon16.png",
-      "32": "assets/icon32.png",
-      "48": "assets/icon48.png",
-      "128": "assets/icon128.png"
-    }
-  },
-  "icons": {
-    "16": "assets/icon16.png",
-    "32": "assets/icon32.png",
-    "48": "assets/icon48.png",
-    "128": "assets/icon128.png"
-  }
-}
