--- chrome-extension-overlay/popup/popup.html
@@ -1,85 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Subtitle Generator</title>
-  <link rel="stylesheet" href="popup.css">
-</head>
-<body>
-  <div class="popup-container">
-    <!-- Header -->
-    <div class="header">
-      <div class="logo-title">
-        <img src="../assets/icon32.png" alt="Logo" class="logo">
-        <h1>Real-time Caption Translator</h1>
-      </div>
-      <button id="themeToggle" class="theme-toggle" title="Toggle theme">
-        <span class="theme-icon">🌙</span>
-      </button>
-    </div>
-    
-    <!-- Status Section -->
-    <div class="status-section">
-      <div class="status-indicator" id="statusIndicator">
-        <div class="status-dot" id="statusDot"></div>
-        <span class="status-text" id="statusText">Ready to start</span>
-      </div>
-    </div>
-    
-    <!-- Controls Section -->
-    <div class="controls-section">
-      <button id="startBtn" class="btn btn-primary">
-        Start Capture
-      </button>
-      
-      <button id="stopBtn" class="btn btn-secondary disabled">
-        Stop Capture
-      </button>
-    </div>
-    
-    <!-- Settings Section -->
-    <div class="settings-section">
-      <div class="setting-group">
-        <label for="targetLanguage">Translate to:</label>
-        <div class="searchable-select">
-          <input type="text" id="languageSearch" class="search-input" placeholder="Search languages...">
-          <div class="dropdown-icon">▼</div>
-          <div id="languageDropdown" class="language-dropdown hidden">
-            <div class="language-option" data-value="none">🚫 No Translation</div>
-          </div>
-        </div>
-        <input type="hidden" id="targetLanguage" value="en">
-      </div>
-
-      <div class="setting-group">
-        <label class="checkbox-label">
-          <input type="checkbox" id="censorProfanity" checked>
-          <span class="checkmark"></span>
-          Censor profanity in Translated caption
-        </label>
-      </div>
-      
-      <div class="setting-group">
-        <label class="checkbox-label">
-          <input type="checkbox" id="showOriginal">
-          <span class="checkmark"></span>
-          Show original caption
-        </label>
-      </div>
-    </div>
-    
-    <!-- Footer -->
-    <div class="footer">
-      <div class="footer-links">
-        <button id="helpBtn" class="link-btn">Help</button>
-      </div>
-      <div class="footer-info">
-        <span>Developed by The Sentinels Team</span>
-      </div>
-    </div>
-  </div>
-  
-  <script src="popup.js"></script>
-</body>
-</html>
