+++ chrome-extension/background/config.js.example
